Add FilterDrawer component for 'refine' filter prototype
